Add suppo for --opa-auth-token-file
Previously kube-mgmt would only load the OPA auth token off the command line. This meant that the token would be possible if users could see the kube-mgmt process in the `ps` output. With this change the auth token can be read out of a file. File loading and command line loading are mutually exclusive.

Signed-off-by: Kevin Fox <Kevin.Fox@pnnl.gov>