Update to client-go release-0.11
This commit updates kube-mgmt to use the most recent stable version of
the client-go library. The client-go library had not been updated in
two years (!!!) so it was a bit overdue given the amount of churn in
client-go.

Theses changes remove support for using kube-mgmt as an initializer as
well as admission controller registration as these features were not
utilized as far as I can tell. The initializer feature was based on
upstream support for initializers which did not exit alpha. The
registration feature was useful for test purposes however in realistic
deployments administrators always manage the webhook configuration
themselves.

These changes also update kube-mgmt to build with Go 1.12. This was
required due to client-go incompatiblity with older stdlib versions.

With these changes, the data synchronizer has been rewritten to use
the new client-go/dynamic client which exposes a much simpler
interface.

Signed-off-by: Torin Sandall <torinsandall@gmail.com>