Move resyncPeriod into the --replicate-resync flag
There is a death spiral in this code if resyncPeriod is too low: if
the number of objects in the API server is too many for OPA to accept
within this time interval, kube-mgmt will endlessly spin on trying to
PUT them all. Hence, this setting must be adjusted to be high enough
for any deployment, and has no safe default other than 0.

I'm leaving the default at 60s for this change, to preserve existing
behaviour.

Signed-off-by: Andrew Suffield <asuffield@gmail.com>