Improve ordering of replication sync operations
Instead of adding all objects twice on startup (once from a List and
again as the cache populates), we now wait for the cache to populate.

syncReset now uploads all objects again within the backoff loop, so
that (a) it will back off if the failure occurs during an object put,
and (b) we no longer depend on resyncPeriod to upload the objects
after a reset, so it is now safe to set it to 0.

Signed-off-by: Andrew Suffield <asuffield@gmail.com>