[charts] fix conditional around deployment annotations.
The conditional on the annotations does not consider annotations users
may have passed via the values, when opa is false and generateCerts is
false the template as it was leads to broken Deployment manifest.

This patch ensures that `opa` and `generateCerts` can be false and not
break the generated Deployment.

Signed-off-by: Àbéjídé Àyodélé <abejideayodele@gmail.com>