Add support for dynamic data synchronization
Given an OPA configuration file, download the bundle, analyze the
references, and automatically setup K8s watches to replicate data into
OPA. This avoids users having to manually configure the replication.
When this mode is enabled, the --health-endpoint option can be used to
enable a /health server on kube-mgmt that can be used for
readiness/liveness checks.

Signed-off-by: Torin Sandall <torin@styra.com>